Update role sharing UI in the application creation wizard
